update BuyButton & new Sabotage subclass
Sabotage subclass serves the purpose of categorizing Powerups for now; should hopefully help with some code later.

Subclasses of Sabotage are all powerups which *only* attack the player's opponent, and does not affect the player themselves.

StealShipment has not been categorized as a sabotage, because while it steals from the opponent, it buffs the player themselves as well, and thus is not purely a sabotage.